+++ crates/cli/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Minor
 
+- Add `--repl` flag to `{applet,platform}-rpc` to enable REPL splitting input as lines
 - Add `platform-clear-store` to clear the store
 - Change `rust-applet-build` to compute the side-table trading applet footprint for performance
 - Add `platform-update` for A/B platform update